requirements.txt
astroid==1.6.0
autopep8==1.3.3
backports.functools-lru-cache==1.4
configparser==3.5.0
enum34==1.1.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.3.1
pylint==1.8.1
pyserial==3.4
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
requests
metawear